+++ packages/orq-rc/RELEASES.md
@@ -7999,3 +7999,13 @@ Based on:
 - [typescript v4.7.0-rc.43] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.43] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.43 - packages/orq-rc
+
+## 2026-04-03 06:53:56
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.44] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.44] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.44 - packages/orq-rc